requirements.txt
Flask
requests
feedparser
beautifulsoup4
python-dotenv